feat: Add PreviewModalContent
