+++ src/app/teams/[id]/panel/widgets/CalenderWidget/CalendarModal.tsx
@@ -1,14 +1,8 @@
 'use client'
-import { ReactNode } from 'react'
-import dayjs from 'dayjs'
-import { Box, IconButton, Stack, Typography } from '@mui/material'
-import CircleIcon from '@mui/icons-material/Circle'
-import { PlusIcon } from '@/icons'
+import { ReactNode, useState } from 'react'
 import CuModal from '@/components/CuModal'
-import useMedia from '@/hook/useMedia'
 import { IEvent } from '@/types/WidgetDataTypes'
-import CalendarLarge from './CalendarLarge'
-import { isPastEvent, useCalendar } from './utils'
+import PreviewModalContent from './PreviewModalContent'
 import * as style from './CalendarModal.style'
 
 interface IPreviewModalProps {
@@ -17,91 +11,41 @@ interface IPreviewModalProps {
   events?: IEvent[]
 }
 
-interface IEventItem {
-  title: ReactNode
-  start: Date | undefined
-  end: Date | undefined
+interface IContainerProps {
+  open: boolean
+  onClose: () => void
+  title: string
+  children: ReactNode
 }
+
+type TCalendarModalType = 'preview' | 'create'
 
 const CalendarModal = ({ open, onClose, events }: IPreviewModalProps) => {
-  const { selectedDate, setSelectedDate, todayEvents, isEmpty } =
-    useCalendar(events)
-  const { isPc } = useMedia()
-  return (
-    <CuModal sx={style.modal} open={open} onClose={onClose} title={'캘린더'}>
-      <Stack
-        sx={isPc ? style.modalContent : style.mobileModalContent}
-        spacing={'2.5rem'}
-      >
-        <CalendarLarge
-          events={events}
-          onDrillDown={(date: Date) => setSelectedDate(dayjs(date))}
-        />
-        <Box>
-          <Stack
-            direction={'row'}
-            alignItems={'center'}
-            justifyContent={'space-between'}
-            sx={style.eventListHeader}
-          >
-            <Typography
-              variant={'Body1Emphasis'}
-              color={'text.strong'}
-            >{`${selectedDate.format('M.D')} 일정`}</Typography>
-            <IconButton>
-              <PlusIcon width={'1.25rem'} height={'1.25rem'} />
-            </IconButton>
-          </Stack>
-          <Stack sx={style.eventList}>
-            {isEmpty ? (
-              <Typography variant={'Body2'} color={'text.alternative'}>
-                등록된 일정이 없습니다.
-              </Typography>
-            ) : (
-              <Stack spacing={'0.5rem'}>
-                {todayEvents?.map((event) => (
-                  <EventItem
-                    key={event.id}
-                    title={event.title}
-                    start={event.start}
-                    end={event.end}
-                  />
-                ))}
-              </Stack>
-            )}
-          </Stack>
-        </Box>
-      </Stack>
-    </CuModal>
-  )
+  const [modalType, setModalType] = useState<TCalendarModalType>('preview')
+  switch (modalType) {
+    case 'preview':
+      return (
+        <CalendarModalContainer open={open} onClose={onClose} title={'캘린더'}>
+          <PreviewModalContent
+            events={events}
+            openNewModal={() => setModalType('create')}
+          />
+        </CalendarModalContainer>
+      )
+    default:
+      return null
+  }
 }
 
-const EventItem = ({ title, start, end }: IEventItem) => {
-  if (!start || !end) return null
-  const isPast = isPastEvent(end)
+const CalendarModalContainer = ({
+  open,
+  onClose,
+  children,
+}: IContainerProps) => {
   return (
-    <Stack direction={'row'} alignItems={'center'} spacing={'0.5rem'}>
-      <CircleIcon
-        sx={{
-          ...style.eventItemIcon,
-          color: isPast ? 'text.assistive' : 'text.strong',
-        }}
-      />
-      <Typography
-        variant={'Body2'}
-        color={isPast ? 'text.assistive' : 'text.strong'}
-        sx={style.eventItemText}
-      >
-        {title}
-      </Typography>
-      <Typography
-        variant={'Caption'}
-        color={isPast ? 'text.assistive' : 'text.alternative'}
-        sx={style.eventItemTime}
-      >
-        {`${dayjs(start).format('HH:mm')}-${dayjs(end).format('HH:mm')}`}
-      </Typography>
-    </Stack>
+    <CuModal sx={style.modal} open={open} onClose={onClose} title={'캘린더'}>
+      <>{children}</>
+    </CuModal>
   )
 }
